Extract Ocaml hazel_semantics to Coq
Used [coq-of-ocaml](https://github.com/clarus/coq-of-ocaml) to extract
Ocaml to Coq as a baseline. See
[paper](https://ocaml.org/meetings/ocaml/2014/ocaml2014_15.pdf)

Currently the extracted Ocaml from this code is garbage, but it does
compile properly in Coq.

Will go through and fixup all the effectful code that was generated so
that the Coq -> Ocaml extraction works properly. At least we know the
data types are consistent with the Ocaml hazel_semantics.ml now.

I wanted to see if Ocaml -> Coq (f) and then Coq -> Ocaml
(g) behaved such that (f o g) is equivalent to (f o g o f o g)
(this means the two extractions would form an adjunction right?) but the
Coq -> Ocaml extraction fails on the second pass.