Minor README/makefile fixes (#27)
* README updates

* now `make clean` won't fail if files are missing

might as well have the `-f` here, since it's already present when deleting
the `_build` folder, and i think it's expected for `make clean` to forcibly
delete things, rather than to be interactive. But the main reason for this
is so that `make clean` doesn't fail when one of the files it's trying to
delete is nonexistant.