Remove concrete syntax for non-empty holes
NEHoles are now being treated purely as an artifact of type-checking -
the `fix_holes` functions will automatically insert them wherever
necessary, the UI makes them look more like annotation than syntax,
and it provides no way to insert them manually. So there really
shouldn't be a way to express NEHoles in the concrete syntax either.
The parser will have no knowledge of NEHoles or the type system, so
it is up to the calling code to use `syn_fix_holes` on the result of
`parse_uhexp` in order to insert the NEHoles appropriately.

This does mean that someone can construct a UHExp with unnecessary
NEHoles, and when they serialize it those unnecessary holes will be
lost, but this isn't a use case that needs to be supported.