Remove metavar for empty holes in concrete syntax, and permit `fix_and_renumber_holes` to overwrite empty hole numbers
In accordance with the previous commit, which removes NEHoles from the
concrete syntax, we should also remove the metavar information from
empty holes in the concrete syntax. But that means that during parsing,
empty holes will have to be provided a dummy metavar.
`fix_and_renumber_holes` provides a way to assign valid metavariables
to the empty holes in a `UHExp` after parsing, and to ensure that these
metavars don't collide with those of NEHoles in the same expression.