Delete barely-used logging to sever heavy dependency
Associator.re had some logging functions that could be leveraged within
Coq. The logging function was only used once and in that case not for
anything that seemed particularly crucial nor particularly informative,
so I think it's ok to delete. Regardless, we may want logging within the
Coq, but I'd think a cleaner solution would be to pass a LOGGER into
Coq, that is completely separate from Associator. The logger could
easily be a stub for builds that don't require JS. By removing the
logging interface from Associator, this commit means that Associator,
which deals with the abstract syntax, doesn't have a dependency on JS.
This dependency was causing some sort of problem during test builds, so
I just severed it.