Serialization now formats things in a vaguely reasonable way
The formatting uses Format's box-style pretty printing, which is easy
enough to use and fairly powerful, but which is very spendthrift with
screen space, which can't deal with unicode characters properly, and
which can't easily conform to some of the more sophisticated formatting
of View.re. In the future we might want to address these shortcomings
but for now, it outputs code that is remotely readable, which is a big
improvement and good enough for the time being.