Don't auto-contextualize test data
It would be useful in the future to be able to open the test data files
as valid Hazel, with some program that can do so. This means that they
should not just be code snippets that expect an unstated context - all
the variables referenced in the test cases should be explicitly defined
in the test data files. This also has the effect of making the test
suite a bit clearer and more explicit, while being a bit more verbose
and also making test cases appear a bit less independent (tho they're
not actually less independent).