Concrete syntax no longer supports non-ascii unicode
The pretty printer used for serializing would indent a bit incorrectly
as a result of unicode characters. Also, on my machine, the unicode
arrows look kind of ugly, overlapping with adjacent characters. For now,
we'll just disable support for unicode characters and the concrete
syntax will only use ascii characters.