Move (de)serialization stuff into main directory
I can't figure out how to get dune to do anything reasonable with
non-trivial directory structure. Flattening the directory structure got
refmt to work properly, and should make it easier to get the tests to
work properly. In the future, we can unflatten the directory structure
if we can figure out how to make dune actually work correctly with it.