Adds a button that loads an example
Feel free to make a revert commit to undo this change, if the example
button is superfluous or ugly - I have given it as an example of how you
can make a button that will load into the UHExp box a UHExp that was
deserialized from the specified string. You can copy and paste the tyxml
for that button, and change the string to be something else that you got
from clicking the `serialize` button. When you copy the serialized stuff
from the console, change any `↵` into spaces. Unfortunately I couldn't
figure out how to make a dropdown in tyxml.