s/LivelitHole/LivelitAp/g , in dynamics, to carry slices to evaluation
Previously, `syn_expand` would convert an `ApLivelit` into a function
application, then wrap it in a `LivelitHole`, which was essentially a
marker that contained no additional semantic info, but contained
metainfo about the livelit and environment for use by the context
inspector. However, to get liveness working properly, each evaluated
`ApLivelit` needs to keep track of the evaluated splices individually,
so we add an additional `LivelitAp` ctor, which carries the expanded
model and the expanded slices (as well as the metainfo about livelit
name, instance, and environment) into evaluation, so that later we can
have evaluation wrap its result in a `LivelitInfo`, which is much like
`LivelitHole` except that it also contains all the evaluated splices.