Fix liveness
Previously I got liveness wrong, and had it sending the environment into
the Livelits view function, which doesn't make any sense. This commit
sets it up properly, so that the view function can't see the
environment, but rather, get access (via the monad it returns) to the
evaluated results of all its splices.