replace `VdomWithSplices` monad with non-monad `splice_getters_to_vdom`
`VdomWithSplices` was defined as a monad, which seemed expedient at the
time. However, `get_splice_div` and `get_splice_value` are "pure" - i.e.
their outputs don't depend on when/how they're called, so the "order" in
which they are called doesn't matter. However, casting `VdomWithSplices`
as a monad forces user code to assume that such things might matter,
leading to more complex and tricky code. To alleviate that, we simply
convert it to a function that accepts `get_splice_div` and
`get_splice_value` as arguments, then returns the vdom node as an
expression closed under those functions. It should now be much easier to
implement `view` for livelits with more than a few splices.