fix bug where livelit ancestor of cursor node has no selected instance
new function `cursor_through_insts` will return metavars for livelit
ancestors of the selected node, in addition to that of the selected
node. `cursor_on_inst` will return only the metavar of the selection, if
appropriate, or that of the selection's livelit parent, should the
selection not have a metavar.