Fixes #350 Allow clicking anywhere on the sidebar border to toggle
Before only the small bottom section of the sidebar allowed clicking
to toggle the side bar state. Now anywhere along the border of the
side pane can be clicked. The cursor has also been changed to the
pointing finger to make it more obvious that this area has an action
associated with clicking.