Fix #356 right sidebar goes off-screen when window isn't wide enough
There's a lot going on with this issue. The first was the code editor
was forced to be a fixed 80ch width which caused a lot of layout problems,
this was removed. The body was changed from flex to grid because that
behaves a bit better in this scenario. The code window was given a min-width
(which is arbitrary can be debated what a better size would be) and the body
was given overflow-x. With overflow-x on the body the side panes behave a bit
better now because if they would force the code area to be too small they push
out the page width instead and add a scroll bar. Overall the layout of the
application should behave a bit better, but there might still be some edge
cases. There were a couple scenarios where overflow wasn't specified and so
the page started vibrating (caught in a layout loop).

Tested with Firefox + Chromium on Linux