Move KeyComboAction.get_details into ActionPanel.generate_panel_body
This was done because KeyComboAction.get_details was only every used by ActionPanel.generate_panel_body. It is now named action_of_combo, and takes less inputs.

It was also changed to return an Action.t, rather than an option(Action.t). If a corresponding EditAction is not found, it fails with an error message mentioning what input caused the error.