collect inference results and pass as func arg to various functions
It's better to calculate the inference results when the Statics.info_map
is calculated. Then pass as function argument when constructing views.
Calculating the width of grouts with type suggestions is broken, because
the inference results are not passed. So width of grout is always 1
instead of width of suggestion :(