feat: functional rough draft for pat match example
Pattern matching now provide an example in the error message for a
missing clause. This example is currently rendered rather simply, and
doesn't support more advanced types such as records. The implementation
tries to enumerate the unmatched possibilities starting at the simplest
value (e.g. for ints it starts at 0, and makes its way up until it finds
a value that is missing).

The current approach for handling tuples is a little hacky, and is again
something that needs some improvement. Tests, better rendering, and
support for more match types needs to be added still.