Signficantly improve handling of inconsistent pats
Invalid patterns are still considered in the pattern building process,
but they are treated as holes in the eyes of pattern generation. To
accomplish this, existing constructors like ctrs and constraints had to
be extended to effectively have some sort of status.

An existing bug relating to wildcards and primitive ctrs was fixed.

This commit also tries to improve code quality in some areas.

A few new tests are also added to have a baseline for the new behavior,
but more will be necessary in the future.