Cycle hole semantics + some inference solver
Added a cycle hole type, but it's missing the special treatment it
needs (custom UI and such).

Filled in some solver cases to the best of my knowledge, as well as
updated the solution type to try to use the existing Hazel types instead
of having to create duplicate definitions. I'm not entirely sure if this
will work yet, but it was a small change for now.

Also I'm not sure if the concept of a solution is needed because multi
holes do exist in Hazel. So TBD I guess